Fixed ordering on radial grid array.
Since increasing layer index corresponds to smaller radii, this array
needs to go from largest to smallest. As the code stands currently the
ordering makes no difference because the radial grid's only role is to
calculate the optical depths, which depend only on the absolute value of
the difference of two neighboring radii.