Reduced # of ALI iterations to 1 since we have full Lambda_star.
With full Lambda_star, doing ALI is equivalent to solving the integral version
of the radiative transfer equation, which requires no iteration but is
expensive. I did it this way just to make sure Lambda_star was being calculated
correctly.

Eventually we'll make the bandwidth of Lambda_star a knob. But as the code
stands currently it won't save any memory to reduce the bandwidth of
Lambda_star because we still call numpy.linalg.solve(), which assumes nothing
about the linear system while it's solving the system of equations. If we
wanted to save memory we would need to call, e.g., a tri-diagonal solver.
However I don't think Numpy comes with one built in.